set up joins

+++ uni_app/controllers/course_controller.py
@@ -1,8 +1,8 @@
-from flask import Blueprint, jsonify, request, render_template, redirect, url_for, current_app
+from flask import Blueprint, jsonify, request, render_template, redirect, url_for, current_app, abort
 from main import db
 from models.courses import Course
 from schemas.course_schema import courses_schema, course_schema
-from flask_login import login_required
+from flask_login import login_required, current_user
 import boto3
 
 
@@ -23,7 +23,6 @@ def get_courses():
         "page_title": "Course Index",
         "courses": courses_schema.dump(Course.query.all())
     }
-    print(data)
     return render_template("course_index.html", page_data=data)
 
 # The POST route endpoint
@@ -31,6 +30,7 @@ def get_courses():
 @login_required
 def create_course():
     new_course=course_schema.load(request.form)
+    new_course.creator = current_user
     db.session.add(new_course)
     db.session.commit()
     return redirect(url_for("courses.get_courses"))
@@ -58,13 +58,15 @@ def get_course(id):
     }
     return render_template("course_detail.html", page_data=data)
 
-# A PUT/PATCH route to update course info
 @courses.route("/courses/<int:id>/", methods=["POST"])
 @login_required
 def update_course(id):
     
     course = Course.query.filter_by(course_id=id)
-   
+    
+    if current_user.id != course.first().creator_id:
+        abort(403, "You do not have permission to alter this course")
+
     updated_fields = course_schema.dump(request.form)
     if updated_fields:
         course.update(updated_fields)
@@ -76,16 +78,31 @@ def update_course(id):
     }
     return render_template("course_detail.html", page_data=data)
 
-# Finally, we round out our CRUD resource with a DELETE method
+@courses.route("/courses/<int:id>/enrol/", methods=["POST"])
+@login_required
+def enrol_in_course(id):
+    course = Course.query.filter_by(course_id=id).first()
+    course.students.append(current_user)
+    db.session.commit()
+    return redirect(url_for('users.user_detail'))
+
+@courses.route("/courses/<int:id>/drop/", methods=["POST"])
+@login_required
+def drop_course(id):
+    course = Course.query.filter_by(course_id=id).first()
+    course.students.remove(current_user)
+    db.session.commit()
+    return redirect(url_for('users.user_detail'))
+
 @courses.route("/courses/<int:id>/delete/", methods=["POST"])
 @login_required
 def delete_course(id):
-    # Can't delete a course that doesn't exist, so get_or_404 here is correct
     course = Course.query.get_or_404(id)
-    # delete the course and commit the transaction
+
+    if current_user.id != course.creator_id:
+        abort(403, "You do not have permission to alter this course")
+    
     db.session.delete(course)
     db.session.commit()
-    # We deleted the row in the database but we still have the python object
-    # since we fetched it before we called session.delete, so we can 
-    # serialize it and return it to the user to show them what they deleted!
+    
     return redirect(url_for("courses.get_courses"))
